# student-api-express/swagger.yaml
openapi: 3.0.3
info:
  title: Student Management REST API
  description: Lab 3 RESTful Web Services implementation for managing Student records using Express.js and Spring Boot.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Express.js Local Server
  - url: http://localhost:8080
    description: Spring Boot Local Server
paths:
  /students:
    get:
      summary: Retrieve all students
      description: Returns a list of all students currently stored in the system.
      responses:
        '200':
          description: List of students retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new student
      description: Validates and creates a new student resource.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
            example:
              name: "Aarav Patel"
              email: "aarav@example.com"
              course: "Computer Science"
              semester: 5
      responses:
        '201':
          description: Student created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid or incomplete request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-08-21T14:30:00.000Z"
                status: 400
                error: "Bad Request"
                message: "Validation failed: 'email' must be a valid email address; 'semester' must be between 1 and 8"

  /students/{id}:
    get:
      summary: Get student by ID
      description: Returns a single student record matching the provided identifier.
      parameters:
        - name: id
          in: path
          required: true
          description: Numerical ID of the student to fetch
          schema:
            type: integer
      responses:
        '200':
          description: Student found and returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student with the given ID does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2026-08-21T14:30:00.000Z"
                status: 404
                error: "Not Found"
                message: "Student with ID 999 not found"

    put:
      summary: Update a student
      description: Replaces or updates the student details matching the provided ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Numerical ID of the student to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
            example:
              name: "Aarav Patel"
              email: "aarav.updated@example.com"
              course: "Information Technology"
              semester: 6
      responses:
        '200':
          description: Student updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Student not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete a student
      description: Removes a student record by ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Numerical ID of the student to delete
          schema:
            type: integer
      responses:
        '200':
          description: Student successfully deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Student with ID 1 deleted successfully"
        '204':
          description: No Content (alternative standard REST response)
        '404':
          description: Student not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    StudentInput:
      type: object
      required:
        - name
        - email
        - course
        - semester
      properties:
        name:
          type: string
          example: "Aarav Patel"
        email:
          type: string
          format: email
          example: "aarav@example.com"
        course:
          type: string
          example: "Computer Science"
        semester:
          type: integer
          minimum: 1
          maximum: 8
          example: 5

    Student:
      type: object
      required:
        - id
        - name
        - email
        - course
        - semester
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Aarav Patel"
        email:
          type: string
          format: email
          example: "aarav@example.com"
        course:
          type: string
          example: "Computer Science"
        semester:
          type: integer
          example: 5

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-08-21T14:30:00.000Z"
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid student payload"
        details:
          type: array
          items:
            type: string